+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All changes to this project will be documented in this file.
 
+## July 14, 2023
+
+- Added diagnostics for lexer
+
 ## July 13, 2023
 
 ## Compiler